func/requirements.txt
--extra-index-url https://backend:<FILL_PASSWORD_HERE>@nexus.sigame.com.br/repository/pypi/simple
etria_logger==0.1.8
Flask==2.1.1
python-decouple==3.6
redis==4.2.2
requests==2.27.1